requirements.txt
# python 3.11
langchain-core==1.2.18
langgraph==1.0.6
langchain-openai==1.1.7
langchain-community==0.4.1
langchain-huggingface==1.2.1
transformers==5.3.0
torch==2.10.0
accelerate==1.13.0
sqlglot==30.0.1

requests>=2.32.0
sentence-transformers>=3.0.0
faiss-cpu==1.13.2